# backend/openapi.yaml
openapi: 3.0.0
info:
  title: Obstacle and Shelter API
  description: API for managing obstacle and shelter data
  version: 1.0.0
...
paths:
  /obstacles:
    get:
      summary: Get all obstacles
      tags:
        - Obstacle
      responses:
        "200":
          description: List of obstacles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListObstacleResponse"
        default:
          description: Error Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-amazon-apigateway-integration:
        credentials:
          Fn::Sub: ${ApiRole.Arn}
        uri:
          Fn::Sub: arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${ObstacleFunction.Arn}/invocations
        passthroughBehavior: when_no_templates
        httpMethod: POST
        type: aws_proxy
    post:
      summary: Create an obstacle
      tags:
        - Obstacle
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateObstacleRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Obstacle"
        default:
          description: Error Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-amazon-apigateway-integration:
        credentials:
          Fn::Sub: ${ApiRole.Arn}
        uri:
          Fn::Sub: arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${ObstacleFunction.Arn}/invocations
        passthroughBehavior: when_no_templates
        httpMethod: POST
        type: aws_proxy
      security:
        - firebase: []
  /obstacles/{id}:
    get:
      summary: Get an obstacle by ID
      tags:
        - Obstacle
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Obstacle found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Obstacle"
        default:
          description: Error Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-amazon-apigateway-integration:
        credentials:
          Fn::Sub: ${ApiRole.Arn}
        uri:
          Fn::Sub: arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${ObstacleFunction.Arn}/invocations
        passthroughBehavior: when_no_templates
        httpMethod: POST
        type: aws_proxy
    put:
      summary: Update an obstacle
      tags:
        - Obstacle
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateObstacleRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Obstacle"
        default:
          description: Error Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-amazon-apigateway-integration:
        credentials:
          Fn::Sub: ${ApiRole.Arn}
        uri:
          Fn::Sub: arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${ObstacleFunction.Arn}/invocations
        passthroughBehavior: when_no_templates
        httpMethod: POST
        type: aws_proxy
      security:
        - firebase: []
    delete:
      summary: Delete an obstacle
      tags:
        - Obstacle
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Empty"
        default:
          description: Error Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-amazon-apigateway-integration:
        credentials:
          Fn::Sub: ${ApiRole.Arn}
        uri:
          Fn::Sub: arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${ObstacleFunction.Arn}/invocations
        passthroughBehavior: when_no_templates
        httpMethod: POST
        type: aws_proxy
      security:
        - firebase: []
  /obstacles/{id}/image-upload:
    post:
      summary: Generate presigned URL for image upload
      tags:
        - Obstacle
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                filename:
                  type: string
      responses:
        "200":
          description: Presigned URL
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:
                    type: string
                  image_s3_key:
                    type: string
        default:
          description: Error Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-amazon-apigateway-integration:
        credentials:
          Fn::Sub: ${ApiRole.Arn}
        uri:
          Fn::Sub: arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${ObstacleFunction.Arn}/invocations
        passthroughBehavior: when_no_templates
        httpMethod: POST
        type: aws_proxy
      security:
        - firebase: []
    options:
      summary: CORS support
      tags:
        - Obstacle
      responses:
        "200":
          description: CORS support
          headers:
            Access-Control-Allow-Headers:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Origin:
              schema:
                type: string
      x-amazon-apigateway-integration:
        type: mock
        requestTemplates:
          application/json: '{"statusCode": 200}'
        responses:
          "200":
            statusCode: 200
            responseParameters:
              method.response.header.Access-Control-Allow-Headers: "'Content-Type,Authorization'"
              method.response.header.Access-Control-Allow-Methods: "'OPTIONS,PUT'"
              method.response.header.Access-Control-Allow-Origin: "'*'"
            responseTemplates:
              application/json: "{}"
  /obstacles/{id}/image:
    put:
      summary: Save image_s3_key to obstacle
      tags:
        - Obstacle
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                image_s3_key:
                  type: string
      responses:
        "200":
          description: Updated obstacle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Obstacle"
        default:
          description: Error Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-amazon-apigateway-integration:
        credentials:
          Fn::Sub: ${ApiRole.Arn}
        uri:
          Fn::Sub: arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${ObstacleFunction.Arn}/invocations
        passthroughBehavior: when_no_templates
        httpMethod: POST
        type: aws_proxy
      security:
        - firebase: []
    options:
      summary: CORS support
      tags:
        - Obstacle
      responses:
        "200":
          description: CORS support
          headers:
            Access-Control-Allow-Headers:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Origin:
              schema:
                type: string
      x-amazon-apigateway-integration:
        type: mock
        requestTemplates:
          application/json: '{"statusCode": 200}'
        responses:
          "200":
            statusCode: 200
            responseParameters:
              method.response.header.Access-Control-Allow-Headers: "'Content-Type,Authorization'"
              method.response.header.Access-Control-Allow-Methods: "'OPTIONS,PUT'"
              method.response.header.Access-Control-Allow-Origin: "'*'"
            responseTemplates:
              application/json: "{}"
  /route-with-obstacles:
    post:
      summary: Get route from Valhalla with obstacles information
      tags:
        - Valhalla
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteWithObstaclesRequest"
      responses:
        "200":
          description: Route with obstacles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValhallaRouteResponse"
        default:
          description: Error Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-amazon-apigateway-integration:
        credentials:
          Fn::Sub: ${ApiRole.Arn}
        uri:
          Fn::Sub: arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${ObstacleFunction.Arn}/invocations
        passthroughBehavior: when_no_templates
        httpMethod: POST
        type: aws_proxy
    options:
      summary: CORS support
      tags:
        - Valhalla
      responses:
        "200":
          description: CORS support
          headers:
            Access-Control-Allow-Headers:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Origin:
              schema:
                type: string
      x-amazon-apigateway-integration:
        type: mock
        requestTemplates:
          application/json: '{"statusCode": 200}'
        responses:
          "200":
            statusCode: 200
            responseParameters:
              method.response.header.Access-Control-Allow-Headers: "'Content-Type,Authorization'"
              method.response.header.Access-Control-Allow-Methods: "'OPTIONS,POST'"
              method.response.header.Access-Control-Allow-Origin: "'*'"
            responseTemplates:
              application/json: "{}"
  /locate:
    post:
      summary: Get nearest road information from Valhalla locate endpoint
      tags:
        - Valhalla
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                locations:
                  type: array
                  items:
                    type: object
                    properties:
                      lat:
                        type: number
                      lon:
                        type: number
                    required:
                      - lat
                      - lon
                  minItems: 1
                costing:
                  type: string
                  default: "pedestrian"
              required:
                - locations
      responses:
        "200":
          description: Nearest road information
          content:
            application/json:
              schema:
                type: object
                description: "Valhalla locate response (structure varies)"
        default:
          description: Error Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-amazon-apigateway-integration:
        credentials:
          Fn::Sub: ${ApiRole.Arn}
        uri:
          Fn::Sub: arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${ObstacleFunction.Arn}/invocations
        passthroughBehavior: when_no_templates
        httpMethod: POST
        type: aws_proxy
    options:
      summary: CORS support
      tags:
        - Valhalla
      responses:
        "200":
          description: CORS support
          headers:
            Access-Control-Allow-Headers:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Origin:
              schema:
                type: string
      x-amazon-apigateway-integration:
        type: mock
        requestTemplates:
          application/json: '{"statusCode": 200}'
        responses:
          "200":
            statusCode: 200
            responseParameters:
              method.response.header.Access-Control-Allow-Headers: "'Content-Type,Authorization'"
              method.response.header.Access-Control-Allow-Methods: "'OPTIONS,POST'"
              method.response.header.Access-Control-Allow-Origin: "'*'"
            responseTemplates:
              application/json: "{}"
  /trace_attributes:
    post:
      summary: Get trace attributes from Valhalla
      tags:
        - Valhalla
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: "Valhalla trace_attributes request body"
      responses:
        "200":
          description: Trace attributes response
          content:
            application/json:
              schema:
                type: object
                description: "Valhalla trace_attributes response"
        default:
          description: Error Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-amazon-apigateway-integration:
        credentials:
          Fn::Sub: ${ApiRole.Arn}
        uri:
          Fn::Sub: arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${ObstacleFunction.Arn}/invocations
        passthroughBehavior: when_no_templates
        httpMethod: POST
        type: aws_proxy
    options:
      summary: CORS support
      tags:
        - Valhalla
      responses:
        "200":
          description: CORS support
          headers:
            Access-Control-Allow-Headers:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Origin:
              schema:
                type: string
      x-amazon-apigateway-integration:
        type: mock
        requestTemplates:
          application/json: '{"statusCode": 200}'
        responses:
          "200":
            statusCode: 200
            responseParameters:
              method.response.header.Access-Control-Allow-Headers: "'Content-Type,Authorization'"
              method.response.header.Access-Control-Allow-Methods: "'OPTIONS,POST'"
              method.response.header.Access-Control-Allow-Origin: "'*'"
            responseTemplates:
              application/json: "{}"
  /trace_route:
    post:
      summary: Get trace route from Valhalla
      tags:
        - Valhalla
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: "Valhalla trace_route request body"
      responses:
        "200":
          description: Trace route response
          content:
            application/json:
              schema:
                type: object
                description: "Valhalla trace_route response"
        default:
          description: Error Response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-amazon-apigateway-integration:
        credentials:
          Fn::Sub: ${ApiRole.Arn}
        uri:
          Fn::Sub: arn:aws:apigateway:${AWS::Region}:lambda:path/2015-03-31/functions/${ObstacleFunction.Arn}/invocations
        passthroughBehavior: when_no_templates
        httpMethod: POST
        type: aws_proxy
    options:
      summary: CORS support
      tags:
        - Valhalla
      responses:
        "200":
          description: CORS support
          headers:
            Access-Control-Allow-Headers:
              schema:
                type: string
            Access-Control-Allow-Methods:
              schema:
                type: string
            Access-Control-Allow-Origin:
              schema:
                type: string
      x-amazon-apigateway-integration:
        type: mock
        requestTemplates:
          application/json: '{"statusCode": 200}'
        responses:
          "200":
            statusCode: 200
            responseParameters:
              method.response.header.Access-Control-Allow-Headers: "'Content-Type,Authorization'"
              method.response.header.Access-Control-Allow-Methods: "'OPTIONS,POST'"
              method.response.header.Access-Control-Allow-Origin: "'*'"
            responseTemplates:
              application/json: "{}"
...
components:
    schemas:
      Error:
        type: object
        properties:
          status_code:
            type: integer
          message:
            type: string
          errors:
            type: object
            additionalProperties:
              type: array
              items:
                type: string
        required:
          - status_code
          - message
...
      Obstacle:
        type: object
        properties:
          id:
            type: integer
          position:
            type: array
            items:
              type: number
            description: "[latitude, longitude]"
            minItems: 2
            maxItems: 2
          type:
            $ref: "#/components/schemas/ObstacleType"
          description:
            type: string
          dangerLevel:
            $ref: "#/components/schemas/DangerLevel"
          nodes:
            type: array
            items:
              type: number
            description: "[latitude, longitude]"
            minItems: 2
            maxItems: 2
          nearestDistance:
            type: number
          noNearbyRoad:
            type: boolean
          createdAt:
            type: string
            format: date-time
          user_id:
            type: string
            description: "作成ユーザーのFirebase UID"
        required:
          - position
          - type
          - description
          - dangerLevel
          - createdAt
          - user_id
      CreateObstacleRequest:
        type: object
        properties:
          position:
            type: array
            items:
              type: number
            description: "[latitude, longitude]"
            minItems: 2
            maxItems: 2
          type:
            $ref: "#/components/schemas/ObstacleType"
          description:
            type: string
          dangerLevel:
            $ref: "#/components/schemas/DangerLevel"
          nodes:
            type: array
            items:
              type: number
            description: "[latitude, longitude]"
            minItems: 2
            maxItems: 2
          nearestDistance:
            type: number
          noNearbyRoad:
            type: boolean
        required:
          - position
          - type
          - dangerLevel
...
      ListObstacleResponse:
        type: object
        properties:
          items:
            type: array
            items:
              $ref: "#/components/schemas/Obstacle"
...
      Empty:
        type: object
...
    securitySchemes:
      firebase:
        type: http
        scheme: bearer
        bearerFormat: JWT
        description: Firebase ID Token